moved store, reducers and store folder to redux
