+++ src/components/Router.js
@@ -4,7 +4,7 @@
 import React from 'react';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import store from '../store/index';
+import store from '../redux/store/index';
 import App from '../App';
 import Recipe from './Recipe';
