+++ src/redux/reducers/categories.js
